+++ packages/atlas/lib/atlas.dart
@@ -3,5 +3,6 @@ library atlas;
 export 'src/atlas_provider.dart';
 export 'src/atlas.dart';
 export 'src/camera_position.dart';
-export 'src/coordinate_pair.dart';
+export 'src/lat_lng.dart';
+export 'src/marker.dart';
 export 'src/provider.dart';