+++ packages/atlas/CHANGELOG.md
@@ -1,5 +1,13 @@
 # atlas
 
+# v0.0.2 (2019-5-17)
+
+Map Marker Integration and Documentation Updates
+
+- Marker Definition and Integration
+- Added Documentation to Atlas APIs
+- Renamed `CoordinatePair` to `LatLng`
+
 # v0.0.1 (2019-5-16)
 
 - Initial Api Definition and Implementation